+++ activesupport/test/log_subscriber_test.rb
@@ -18,7 +18,11 @@ class MyLogSubscriber < ActiveSupport::LogSubscriber
   end
 
   def bar(event)
-    info "#{color("cool", :red)}, #{color("isn't it?", :blue, true)}"
+    info "#{color("cool", :red)}, #{color("isn't it?", :blue, bold: true)}"
+  end
+
+  def baz(event)
+    info "#{color("rad", :green, bold: true, underline: true)}, #{color("isn't it?", :yellow, italic: true)}"
   end
 
   def puke(event)
@@ -54,6 +58,12 @@ class SyncLogSubscriberTest < ActiveSupport::TestCase
     ActiveSupport::LogSubscriber.colorize_logging = true
     @log_subscriber.bar(nil)
     assert_equal "\e[31mcool\e[0m, \e[1m\e[34misn't it?\e[0m", @logger.logged(:info).last
+  end
+
+  def test_set_mode_for_messages
+    ActiveSupport::LogSubscriber.colorize_logging = true
+    @log_subscriber.baz(nil)
+    assert_equal "\e[31rad\e[0m, \e[1m\e[34misn't it?\e[0m", @logger.logged(:info).last
   end
 
   def test_does_not_set_color_if_colorize_logging_is_set_to_false